+++ front/config.php
@@ -40,10 +40,7 @@
  **/
 
 // https://codeberg.org/QuinQuies/glpisaml/issues/73
-// use Html;
-// use Search;
-// use Plugin;
-use GlpiPlugin\Glpisaml\Config as samlConfig;
+use GlpiPlugin\Samlsso\Config as samlConfig;
 
 include_once '../../../inc/includes.php';               //NOSONAR - Cannot be included with USE keyword
 // Check the rights