+++ front/exclude.php
@@ -42,7 +42,7 @@
 use Html;
 use Plugin;
 use Search;
-use GlpiPlugin\Glpisaml\Exclude;
+use GlpiPlugin\Samlsso\Exclude;
 
 include "../../../inc/includes.php";                    //NOSONAR - Cant be included with USE.
 // Check the rights