+++ setup.php
@@ -43,27 +43,27 @@
 // This file is included into the GLPI Plugin base class.
 // https://codeberg.org/QuinQuies/glpisaml/issues/18
 use Glpi\Plugin\Hooks;
-use GlpiPlugin\Glpisaml\Config;
-use GlpiPlugin\Glpisaml\Exclude;
-use GlpiPlugin\Glpisaml\LoginFlow;
-use GlpiPlugin\Glpisaml\RuleSamlCollection;
+use GlpiPlugin\Samlsso\Config;
+use GlpiPlugin\Samlsso\Exclude;
+use GlpiPlugin\Samlsso\LoginFlow;
+use GlpiPlugin\Samlsso\RuleSamlCollection;
 
 // Setup constants
-define('PLUGIN_GLPISAML_VERSION', '1.1.11');
-define('PLUGIN_GLPISAML_MIN_GLPI', '10.0.11');                                                  // Min required GLPI version
-define('PLUGIN_GLPISAML_MAX_GLPI', '10.9.99');                                                  // Max GLPI compat version
-define('PLUGIN_NAME', 'glpisaml');                                                              // Plugin name
-define('LOGEVENTS','events');                                                                   // specifies log extention
-define('LOGERRORS','errors');                                                                   // specifies log extention
+define('PLUGIN_SAMLSSO_VERSION', '1.2.00');
+define('PLUGIN_SAMLSSO_MIN_GLPI', '10.0.11');                                                  // Min required GLPI version
+define('PLUGIN_SAMLSSO_MAX_GLPI', '11.9.00');                                                  // Max GLPI compat version
+define('PLUGIN_NAME', 'samlsso');                                                              // Plugin name
+define('LOGEVENTS','events');                                                                  // specifies log extention
+define('LOGERRORS','errors');                                                                  // specifies log extention
 // Directories
-define('PLUGIN_GLPISAML_WEBDIR', Plugin::getWebDir(PLUGIN_NAME, false));                        // Plugin web directory
-define('PLUGIN_GLPISAML_SRCDIR', __DIR__ . '/src');                                             // Location of the main classes
+define('PLUGIN_SAMLSSO_WEBDIR', Plugin::getWebDir(PLUGIN_NAME, false));                        // Plugin web directory
+define('PLUGIN_SAMLSSO_SRCDIR', __DIR__ . '/src');                                             // Location of the main classes
 // WebPaths
-define('PLUGIN_GLPISAML_ATOM_URL', 'https://codeberg.org/QuinQuies/glpisaml/releases.rss');     // Location of the repository versions
-define('PLUGIN_GLPISAML_META_PATH', '/front/meta.php');                                         // Location where to get metadata about sp
-define('PLUGIN_GLPISAML_CONF_PATH', '/front/config.php');                                       // Location of the config page
-define('PLUGIN_GLPISAML_CONF_FORM', '/front/config.form.php');                                  // Location of config form
-define('PLUGIN_GLPISAML_CONFCSS_PATH', 'templates/css/glpiSaml.css');                           // Location of the config CSS
+// define('PLUGIN_SAMLSSO_ATOM_URL', 'https://github.com/DonutsNL/samlSSO/releases.rss');         // Location of the repository versions
+define('PLUGIN_SAMLSSO_META_PATH', '/front/meta.php');                                         // Location where to get metadata about sp
+define('PLUGIN_SAMLSSO_CONF_PATH', '/front/config.php');                                       // Location of the config page
+define('PLUGIN_SAMLSSO_CONF_FORM', '/front/config.form.php');                                  // Location of config form
+define('PLUGIN_SAMLSSO_CONFCSS_PATH', 'templates/css/samlSSO.css');                            // Location of the config CSS
 
 /**
  * Default GLPI Plugin Init function.
@@ -71,16 +71,16 @@ define('PLUGIN_GLPISAML_CONFCSS_PATH', 'templates/css/glpiSaml.css');           
  * @see https://glpi-developer-documentation.readthedocs.io/en/master/plugins/requirements.html
  * @see https://codeberg.org/QuinQuies/glpisaml/issues/8
  */
-function plugin_init_glpisaml() : void                                                          //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
+function plugin_init_samlsso() : void                                                          //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
 {
-    global $PLUGIN_HOOKS;                                                                       //NOSONAR
+    global $PLUGIN_HOOKS;                                                                      //NOSONAR
     $plugin = new Plugin();
 
     // INCLUDE PSR4 AUTOLOADER
-    include_once(__DIR__. '/vendor/autoload.php');                                              //NOSONAR - intentional include_once to load composer autoload;
+    include_once(__DIR__. '/vendor/autoload.php');                                             //NOSONAR - intentional include_once to load composer autoload;
 
     // CSRF
-    $PLUGIN_HOOKS[Hooks::CSRF_COMPLIANT][PLUGIN_NAME] = true;                                   //NOSONAR - These are GLPI default variable names  
+    $PLUGIN_HOOKS[Hooks::CSRF_COMPLIANT][PLUGIN_NAME] = true;                                  //NOSONAR - These are GLPI default variable names  
     
     // Do not show config buttons if plugin is not enabled.
     if ($plugin->isInstalled(PLUGIN_NAME) || $plugin->isActivated(PLUGIN_NAME)) {
@@ -91,10 +91,10 @@ function plugin_init_glpisaml() : void                                          
 
         // Hook the configuration page
         if (Session::haveRight('config', UPDATE)) {
-            $PLUGIN_HOOKS['config_page'][PLUGIN_NAME]       = PLUGIN_GLPISAML_CONF_PATH;        //NOSONAR
+            $PLUGIN_HOOKS['config_page'][PLUGIN_NAME]       = PLUGIN_SAMLSSO_CONF_PATH;        //NOSONAR
         }
         $PLUGIN_HOOKS['menu_toadd'][PLUGIN_NAME]['config']  = [Config::class];
-        $PLUGIN_HOOKS[Hooks::ADD_CSS][PLUGIN_NAME][]        = PLUGIN_GLPISAML_CONFCSS_PATH;
+        $PLUGIN_HOOKS[Hooks::ADD_CSS][PLUGIN_NAME][]        = PLUGIN_SAMLSSO_CONFCSS_PATH;
 
         // Register and hook the saml rules
         Plugin::registerClass(RuleSamlCollection::class, ['rulecollections_types' => true]);
@@ -102,10 +102,10 @@ function plugin_init_glpisaml() : void                                          
 
         // Register and hook the loginFlow directly after GLPI init.
         Plugin::registerClass(LoginFlow::class);
-        $PLUGIN_HOOKS[Hooks::POST_INIT][PLUGIN_NAME]        = 'plugin_glpisaml_evalAuth';       //NOSONAR
+        $PLUGIN_HOOKS[Hooks::POST_INIT][PLUGIN_NAME]        = 'plugin_samlsso_evalAuth';       //NOSONAR
 
         // Hook the login buttons
-        $PLUGIN_HOOKS[Hooks::DISPLAY_LOGIN][PLUGIN_NAME]    = 'plugin_glpisaml_displaylogin';
+        $PLUGIN_HOOKS[Hooks::DISPLAY_LOGIN][PLUGIN_NAME]    = 'plugin_samlsso_displaylogin';
     }
 }
 
@@ -114,18 +114,18 @@ function plugin_init_glpisaml() : void                                          
  * Returns the name and the version of the plugin
  * @return array
  */
-function plugin_version_glpisaml() : array                                                      //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
+function plugin_version_samlsso() : array                                                      //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
 {
     return [
         'name'           => 'samlSSO',
-        'version'        => PLUGIN_GLPISAML_VERSION,
+        'version'        => PLUGIN_SAMLSSO_VERSION,
         'author'         => 'Chris Gralike',
-        'license'        => 'GPLv2+',
-        'homepage'       => 'https://codeberg.org/QuinQuies/glpisaml',
+        'license'        => 'GPLv3',
+        'homepage'       => 'https://github.com/DonutsNL/samlSSO/',
         'requirements'   => [
             'glpi' => [
-            'min' => PLUGIN_GLPISAML_MIN_GLPI,
-            'max' => PLUGIN_GLPISAML_MAX_GLPI,
+            'min' => PLUGIN_SAMLSSO_MIN_GLPI,
+            'max' => PLUGIN_SAMLSSO_MAX_GLPI,
             ],
             'php'    => [
             'min' => '8.0'
@@ -139,7 +139,7 @@ function plugin_version_glpisaml() : array                                      
  * Check pre-requisites before install
  * @return boolean
  */
-function plugin_glpisaml_check_prerequisites() : bool                                           //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
+function plugin_samlsso_check_prerequisites() : bool                                           //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
 {
     // include plugin composer
     // https://github.com/pluginsGLPI/example/issues/49#issuecomment-1891552141
@@ -162,7 +162,7 @@ function plugin_glpisaml_check_prerequisites() : bool                           
  * @param boolean $verbose Whether to display message on failure. Defaults to false
  * @return boolean
  */
-function plugin_glpisaml_check_config($verbose = false) : bool                                  //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
+function plugin_samlsso_check_config($verbose = false) : bool                                  //NOSONAR - phpcs:ignore PSR1.Function.CamelCapsMethodName
 {
    if ($verbose) {
       echo __('Installed / not configured', PLUGIN_NAME);